fix(relocation): keep i18n barrel pure
